refactor(client): rename ProblemCategoryPage components to follow conventions
